Fixes types in ArrayHelper.zip, removes unnecessary variables
