+++ src/assets/styles/styles/custom/_variables.scss
@@ -22,24 +22,9 @@ $dzien-noc: #641ea9;
 $rano: rgba(255, 209, 0, 1);
 $popoludnie: rgba(0, 163, 255, 1);
 $noc: rgba(16, 32, 70, 1);
-
 $l4: rgba(198, 0, 0, 1);
-$l4-50: change-color(
-  $color: $l4,
-  $alpha: 0.3,
-);
-
 $kwaratanna: rgba(0, 0, 0, 1);
-$kwaratanna-50: change-color(
-  $color: $kwaratanna,
-  $alpha: 0.3,
-);
-
 $urlop: rgba(255, 138, 0, 1);
-$urlop-50: change-color(
-  $color: $urlop,
-  $alpha: 0.3,
-);
 
 // Color system
 $primary: #1d3557 !default;
