+++ src/assets/styles/styles/custom/_workers-calendar.scss
@@ -85,8 +85,8 @@
         width: 0;
       }
 
-      background-color: $l4-50;
-      color: $l4-50;
+      background-color: change-color($color: $l4, $alpha: 0.3);
+      color: change-color($color: $l4, $alpha: 0.3);
     }
 
     .keepOntrueK {
@@ -98,8 +98,8 @@
         width: 0;
       }
 
-      background-color: $kwaratanna-50;
-      color: $kwaratanna-50;
+      background-color: change-color($color: $kwaratanna, $alpha: 0.3);
+      color: change-color($color: $kwaratanna, $alpha: 0.3);
     }
 
     .keepOntrueU {
@@ -111,7 +111,7 @@
         width: 0;
       }
 
-      background-color: $urlop-50;
+      background-color: change-color($color: $urlop, $alpha: 0.3);
     }
 
     .hasNextfalse {
@@ -145,7 +145,7 @@
         background-color: $l4;
       }
 
-      background-color: $l4-50;
+      background-color: change-color($color: $l4, $alpha: 0.3);
       color: $white;
     }
 
@@ -154,7 +154,7 @@
         background-color: $kwaratanna;
       }
 
-      background-color: $kwaratanna-50;
+      background-color: change-color($color: $kwaratanna, $alpha: 0.3);
       color: $white;
     }
 
@@ -163,7 +163,7 @@
         background-color: $urlop;
       }
 
-      background-color: $urlop-50;
+      background-color: change-color($color: $urlop, $alpha: 0.3);
     }
   }
 }
